polymarket-copy-trading/telegram-bot/requirements.txt
python-telegram-bot==20.7
httpx==0.26.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
python-dotenv==1.0.0
PyJWT==2.8.0
